Chats Tab is officially connected to Firebase
I can't do much because I would need discuss it with the others so yea. But hopfully the plan is to make it where when you add someone they are automatically added in your chat list. Thats the idea. I should try to add a search bar for specific chats too. But for now lets keep it simple